django_api/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
psycopg2-binary
django-health-check
psutil
requests
